prediction binary input move math in a submodule
