+++ docs/source/api.rst
@@ -113,6 +113,7 @@ Binary nodes
     prediction_mean_value_parent
     prediction_precision_value_parent
     prediction_value_parent
+    prediction_input_value_parent
 
 
 Distribution
